♻️ Replace vanniktech BlurHash with pure-Kotlin decoder
Extract decode algorithm to BlurHashCore in commonMain. Android and
desktop actuals now use shared core with platform-specific bitmap
conversion. Removes com.vanniktech:blurhash dependency entirely.
