+++ gradle/libs.versions.toml
@@ -31,7 +31,6 @@ coil = "3.3.0"
 media3 = "1.9.2"
 concurrent-futures = "1.3.0"
 markdown-renderer = "0.39.2"
-blurhash = "0.3.0"
 kmpalette = "3.1.0"
 
 # Swift Interop
@@ -133,8 +132,6 @@ media3-datasource-okhttp = { module = "androidx.media3:media3-datasource-okhttp"
 concurrent-futures = { module = "androidx.concurrent:concurrent-futures-ktx", version.ref = "concurrent-futures" }
 # Markdown Rendering
 markdown-renderer-m3 = { module = "com.mikepenz:multiplatform-markdown-renderer-m3", version.ref = "markdown-renderer" }
-# BlurHash (image placeholders)
-blurhash = { module = "com.vanniktech:blurhash", version.ref = "blurhash" }
 # KMPalette (cross-platform color extraction from images)
 kmpalette-core = { module = "com.kmpalette:kmpalette-core", version.ref = "kmpalette" }
